Created content for oatmeal recipe page
Added an image of oatmeal.
Added a subheading named description.
Added a paragraph that describes the food.
Added a subheading named ingredients.
Added an unordered list that details the ingredients needed to make
oatmeal.
Added a subheading named steps.
Added a ordered list that details the steps needed to make oatmeal.